Add WarningChecksPublisher
